Sprint 710: 2026-04-20 refactor pass (upload unification + helpers split + route thinning)
Retroactive — body written 2026-04-20, never committed. Sprint 693
(already pushed as d22dc86) imports uploadTrialBalance from the Priority-1
module below; without this commit, a fresh checkout fails next build on
an unresolved module. Full scope: docs/architecture/REFACTOR_2026_04_20.md

No behaviour change. No API contract change. Additive shim-preserving
refactor — every `from shared.helpers import X` call site keeps working.

Priority 1 — frontend TB upload unification:
  new utils/trialBalanceUpload.ts — buildTrialBalanceFormData +
  uploadTrialBalance with discriminated outcome union (success /
  audit_error / auth / email_not_verified / forbidden / transport).
  15-test suite. multi-period/page.tsx migrates off ad-hoc apiPost and
  gains the same EMAIL_NOT_VERIFIED / 403 discrimination as the main
  TB flow.

Priority 2 — shared/helpers.py decomposition (1075 -> ~150 line shim):
  new upload_pipeline.py (ten-step upload defense), filenames.py
  (hash_filename / safe_download_filename / sanitize_csv_value),
  background_email.py (safe_background_email wrapper), tool_run_recorder.py
  (maybe_record_tool_run + _log_tool_activity).

Priority 3 — route thinning:
  new organization_policy.py (slugify / unique_slug / require_admin|owner
  / apply_org_subscription_to_user / revert_user_tier_after_removal),
  billing_access.py (require_billing_analytics_access).
  routes/organization.py -60 lines, routes/billing.py -30 lines.

Tests verified 2026-04-21: test_refactor_2026_04_20 (13), backend
regression across org/billing/upload/export_helpers (303), frontend
trialBalanceUpload + useTrialBalanceUpload + useAuditUpload +
uploadTransport + MultiPeriod (53). All pass, no regressions.
